Uses different collision templates for different ability types
