+++ Danki2/Assets/Scripts/Abilities2/AbilityCastInformation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d90edef0d8d0af54dad08d52636d4719
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
